services/transaction_service/requirements.txt
django
gunicorn
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
ldap3
drf-spectacular>=0.28.0
drf-spectacular-sidecar>=2025.9.1
python-dotenv